requirements.txt
dbt-snowflake==1.3.0rc2
snowflake-snowpark-python==0.11.0